+++ src/main/java/in/jam/runner/AppRunner.java
@@ -19,9 +19,9 @@ private EligibilityDetailsRepo repo;
 		EligibilityDetails	entity1= new EligibilityDetails();
 		entity1.setEligId(1);
 		entity1.setName("teju");
-		entity1.setMobile((long) 123456789);
+		entity1.setMobile(123456789);
 		entity1.setGender('F');
-		entity1.setSsn((long) 12344);
+		entity1.setSsn(12344);
 		entity1.setPlanName("SNAP");
 		entity1.setPlanStatus("Approved");
 		
@@ -31,9 +31,9 @@ private EligibilityDetailsRepo repo;
 		EligibilityDetails	entity2= new EligibilityDetails();
 		entity2.setEligId(2);
 		entity2.setName("mahi");
-		entity2.setMobile((long) 1234589);
+		entity2.setMobile(1234589);
 		entity2.setGender('M');
-		entity2.setSsn((long) 1244);
+		entity2.setSsn(1244);
 		entity2.setPlanName("CCP");
 		entity2.setPlanStatus("Denied");
 		
@@ -43,9 +43,9 @@ private EligibilityDetailsRepo repo;
 		EligibilityDetails	entity3= new EligibilityDetails();
 		entity3.setEligId(3);
 		entity3.setName("sai");
-		entity3.setMobile((long) 5234589);
+		entity3.setMobile(5234589);
 		entity3.setGender('M');
-		entity3.setSsn((long) 1246);
+		entity3.setSsn(1246);
 		entity3.setPlanName("JAVA");
 		entity3.setPlanStatus("closed");
